+++ mod-source-record-manager-server/src/main/java/org/folio/dao/JobExecutionDaoImpl.java
@@ -18,6 +18,7 @@ import org.folio.rest.jaxrs.model.DeleteJobExecutionsResp;
 import org.folio.rest.jaxrs.model.JobExecution;
 import org.folio.rest.jaxrs.model.JobExecutionDetail;
 import org.folio.rest.jaxrs.model.JobExecutionDto;
+import org.folio.rest.jaxrs.model.JobProfileInfoCollection;
 import org.folio.rest.jaxrs.model.JobExecutionDtoCollection;
 import org.folio.rest.jaxrs.model.JobExecutionUserInfo;
 import org.folio.rest.jaxrs.model.JobExecutionUserInfoCollection;
@@ -39,6 +40,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 import static java.lang.String.format;
@@ -75,6 +77,7 @@ import static org.folio.dao.util.JobExecutionDBConstants.UI_STATUS_FIELD;
 import static org.folio.dao.util.JobExecutionDBConstants.UPDATE_BY_IDS_SQL;
 import static org.folio.dao.util.JobExecutionDBConstants.UPDATE_SQL;
 import static org.folio.dao.util.JobExecutionDBConstants.USER_ID_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.GET_RELATED_JOB_PROFILES_SQL;
 import static org.folio.rest.persist.PostgresClient.convertToPsqlStandard;
 
 /**
@@ -160,6 +163,20 @@ public class JobExecutionDaoImpl implements JobExecutionDao {
   }
 
   @Override
+  public Future<JobProfileInfoCollection> getRelatedJobProfiles(int offset, int limit, String tenantId) {
+    Promise<RowSet<Row>> promise = Promise.promise();
+    try {
+      String jobTable = formatFullTableName(tenantId, TABLE_NAME);
+      String query = format(GET_RELATED_JOB_PROFILES_SQL, jobTable);
+      pgClientFactory.createInstance(tenantId).select(query, Tuple.of(limit, offset), promise);
+    } catch (Exception e) {
+      LOGGER.error("Error getting related Job Profiles", e);
+      promise.fail(e);
+    }
+    return promise.future().map(this::mapRowToJobProfileInfoCollection);
+  }
+
+  @Override
   public Future<String> save(JobExecution jobExecution, String tenantId) {
     Promise<RowSet<Row>> promise = Promise.promise();
     String preparedQuery = String.format(GET_JOB_EXECUTION_HR_ID, PostgresClient.convertToPsqlStandard(tenantId));
@@ -202,7 +219,7 @@ public class JobExecutionDaoImpl implements JobExecutionDao {
         pgClientFactory.createInstance(tenantId).startTx(connection);
         return connection.future();
       }).compose(v -> {
-        String selectForUpdate = format("SELECT * FROM %s WHERE id = $1 LIMIT 1 FOR UPDATE", formatFullTableName(tenantId, TABLE_NAME));
+        String selectForUpdate = format("SELECT * FROM %s WHERE id = $1 AND is_deleted = false LIMIT 1 FOR UPDATE", formatFullTableName(tenantId, TABLE_NAME));
         Promise<RowSet<Row>> selectResult = Promise.promise();
         pgClientFactory.createInstance(tenantId).execute(connection.future(), selectForUpdate, Tuple.of(jobExecutionId), selectResult);
         return selectResult.future();
@@ -323,10 +340,19 @@ public class JobExecutionDaoImpl implements JobExecutionDao {
 
   private JobExecutionDtoCollection mapToJobExecutionDtoCollection(RowSet<Row> rowSet) {
     JobExecutionDtoCollection jobCollection = new JobExecutionDtoCollection().withTotalRecords(0);
-    for (Row row : rowSet) {
+    rowSet.iterator().forEachRemaining(row -> {
       jobCollection.getJobExecutions().add(mapRowToJobExecutionDto(row));
       jobCollection.setTotalRecords(row.getInteger(TOTAL_COUNT_FIELD));
-    }
+    });
+    return jobCollection;
+  }
+
+  private JobProfileInfoCollection mapRowToJobProfileInfoCollection(RowSet<Row> rowSet) {
+    JobProfileInfoCollection jobCollection = new JobProfileInfoCollection().withTotalRecords(0);
+    rowSet.iterator().forEachRemaining(row -> {
+      jobCollection.getJobProfilesInfo().add(mapRowToJobProfileInfo(row));
+      jobCollection.setTotalRecords(row.getInteger(TOTAL_COUNT_FIELD));
+    });
     return jobCollection;
   }
 
@@ -402,12 +428,12 @@ public class JobExecutionDaoImpl implements JobExecutionDao {
 
   private JobProfileInfo mapRowToJobProfileInfo(Row row) {
     UUID profileId = row.getUUID(JOB_PROFILE_ID_FIELD);
-    if (profileId != null) {
+    if (Objects.nonNull(profileId)) {
       return new JobProfileInfo()
         .withId(profileId.toString())
         .withName(row.getString(JOB_PROFILE_NAME_FIELD))
         .withHidden(row.getBoolean(JOB_PROFILE_HIDDEN_FIELD))
-        .withDataType(row.getString(JOB_PROFILE_DATA_TYPE_FIELD) == null
+        .withDataType(Objects.isNull(row.getString(JOB_PROFILE_DATA_TYPE_FIELD))
           ? null : JobProfileInfo.DataType.fromValue(row.getString(JOB_PROFILE_DATA_TYPE_FIELD)));
     }
     return null;