+++ mod-source-record-manager-server/src/main/java/org/folio/dao/util/JobExecutionDBConstants.java
@@ -26,8 +26,9 @@ public final class JobExecutionDBConstants {
   public static final String TOTAL_COUNT_FIELD = "total_count";
   public static final String CURRENTLY_PROCESSED_FIELD = "currently_processed";
   public static final String TOTAL_FIELD = "total";
+  public static final String IS_DELETED_FIELD = "is_deleted";
 
-  public static final String GET_BY_ID_SQL = "SELECT * FROM %s WHERE id = $1";
+  public static final String GET_BY_ID_SQL = "SELECT * FROM %s WHERE id = $1 AND is_deleted = false";
   public static final String UPDATE_BY_IDS_SQL = "UPDATE ${tenantName}.${tableName} SET ${setFieldName} = ${setFieldValue} WHERE ${setConditionalFieldName} IN ('${setConditionalFieldValues}') RETURNING ${returningFieldNames}";
 
   public static final String INSERT_SQL =
@@ -48,13 +49,13 @@ public final class JobExecutionDBConstants {
 
   public static final String GET_CHILDREN_JOBS_BY_PARENT_ID_SQL =
     "WITH cte AS (SELECT count(*) AS total_count FROM %s " +
-    "WHERE parent_job_id = $1 AND subordination_type = 'CHILD') " +
+    "WHERE parent_job_id = $1 AND subordination_type = 'CHILD' AND is_deleted = false) " +
     "SELECT j.*, cte.*, p.total_records_count total, " +
     "p.succeeded_records_count + p.error_records_count currently_processed " +
     "FROM %s j " +
     "LEFT JOIN %s p ON  j.id = p.job_execution_id " +
     "LEFT JOIN cte ON true " +
-    "WHERE parent_job_id = $1 AND subordination_type = 'CHILD' " +
+    "WHERE parent_job_id = $1 AND subordination_type = 'CHILD' AND is_deleted = false " +
     "LIMIT $2 OFFSET $3";
 
   public static final String GET_JOBS_NOT_PARENT_SQL =
@@ -67,6 +68,16 @@ public final class JobExecutionDBConstants {
     "LEFT JOIN cte ON true " +
     "WHERE subordination_type <> 'PARENT_MULTIPLE' AND %s " +
     "%s " +
+    "LIMIT $1 OFFSET $2";
+
+  public static final String GET_RELATED_JOB_PROFILES_SQL =
+    "WITH unique_profiles AS (SELECT DISTINCT job_profile_id, job_profile_name, job_profile_data_type, job_profile_hidden " +
+    "FROM %s " +
+    "WHERE job_profile_id IS NOT NULL AND job_profile_hidden = false), " +
+    "total AS (SELECT count(*) AS total_count FROM unique_profiles) " +
+    "SELECT j.*, p.* " +
+    "FROM unique_profiles j " +
+    "LEFT JOIN total p ON true " +
     "LIMIT $1 OFFSET $2";
 
   public static final String GET_UNIQUE_USERS = "WITH unique_users AS (SELECT DISTINCT user_id, " +